banana/requirements.txt
flask==3.0.0
ultralytics==8.0.0
opencv-python-headless==4.8.0.74
pillow==10.0.0
werkzeug==3.0.0
gunicorn==21.2.0
